--- ui/app/pages/send/send-content/add-recipient/add-recipient.selectors.js
@@ -1,11 +1,0 @@
-export function getToDropdownOpen (state) {
-  return state.send.toDropdownOpen
-}
-
-export function sendToIsInError (state) {
-  return Boolean(state.send.errors.to)
-}
-
-export function getTokens (state) {
-  return state.metamask.tokens
-}
