--- ui/app/pages/send/send-content/send-gas-row/send-gas-row.selectors.js
@@ -1,11 +1,0 @@
-export function getGasLoadingError (state) {
-  return state.send.errors.gasLoading
-}
-
-export function gasFeeIsInError (state) {
-  return Boolean(state.send.errors.gasFee)
-}
-
-export function getGasButtonGroupShown (state) {
-  return state.send.gasButtonGroupShown
-}
