+++ ui/app/pages/send/send.selectors.js
@@ -196,3 +196,19 @@ export function getUnapprovedTxs (state) {
 export function getQrCodeData (state) {
   return state.appState.qrCodeData
 }
+
+export function getGasLoadingError (state) {
+  return state.send.errors.gasLoading
+}
+
+export function gasFeeIsInError (state) {
+  return Boolean(state.send.errors.gasFee)
+}
+
+export function getGasButtonGroupShown (state) {
+  return state.send.gasButtonGroupShown
+}
+
+export function getTokens (state) {
+  return state.metamask.tokens
+}
