+++ ui/app/pages/send/tests/send-selectors.test.js
@@ -36,6 +36,10 @@ import {
   getTokenBalance,
   getTokenExchangeRate,
   getUnapprovedTxs,
+  gasFeeIsInError,
+  getGasLoadingError,
+  getGasButtonGroupShown,
+  getTokens,
 } from '../send.selectors.js'
 import mockState from './send-selectors-test-data'
 
@@ -547,6 +551,75 @@ describe('send selectors', function () {
         }
 
         assert.equal(sendAmountIsInError(state), false)
+      })
+    })
+  })
+
+  describe('send-gas-row selectors', function () {
+
+    describe('getGasLoadingError()', function () {
+      it('should return send.errors.gasLoading', function () {
+        const state = {
+          send: {
+            errors: {
+              gasLoading: 'abc',
+            },
+          },
+        }
+
+        assert.equal(getGasLoadingError(state), 'abc')
+      })
+    })
+
+    describe('gasFeeIsInError()', function () {
+      it('should return true if send.errors.gasFee is truthy', function () {
+        const state = {
+          send: {
+            errors: {
+              gasFee: 'def',
+            },
+          },
+        }
+
+        assert.equal(gasFeeIsInError(state), true)
+      })
+
+      it('should return false send.errors.gasFee is falsely', function () {
+        const state = {
+          send: {
+            errors: {
+              gasFee: null,
+            },
+          },
+        }
+
+        assert.equal(gasFeeIsInError(state), false)
+      })
+    })
+
+    describe('getGasButtonGroupShown()', function () {
+      it('should return send.gasButtonGroupShown', function () {
+        const state = {
+          send: {
+            gasButtonGroupShown: 'foobar',
+          },
+        }
+
+        assert.equal(getGasButtonGroupShown(state), 'foobar')
+      })
+    })
+  })
+
+  describe('add-recipient selectors', function () {
+    describe('getTokens()', function () {
+      it('should return empty array if no tokens in state', function () {
+        const state = {
+          metamask: {
+            tokens: [],
+          },
+        }
+
+        assert.deepStrictEqual(getTokens(state), [])
       })
     })
   })
